requirements.txt
datasets
loralib
sentencepiece
git+https://github.com/huggingface/transformers.git
accelerate
bitsandbytes
git+https://github.com/huggingface/peft.git
gradio
appdirs
markdown2
pygments
mdtex2html
llama_index
openai
html2text